+++ epub.go
@@ -2,6 +2,7 @@
 package epub
 
 import (
+	"archive/zip"
 	"fmt"
 	"os"
 )
@@ -107,19 +108,13 @@ func GetMetadataFromFile(path string) (*MetaInformation, error) {
 // GetPackage reads an epub's Open Package Document from an epub opened as a ReadAtSeeker.
 // Deprecated: interacting with epub through ReadAtSeeker will be suppressed in
 // next version, seems not useful in practice.
-func GetPackage(rzip ReadAtSeeker) (*PackageDocument, error) {
-	c, err := getContainer(rzip)
+func GetPackage(r ReadAtSeeker) (*PackageDocument, error) {
+	zr, err := newZipReader(r)
 	if err != nil {
 		return nil, err
 	}
 
-	r, err := openFromZip(rzip, c.Rootfiles.FullPath)
-	if err != nil {
-		return nil, err
-	}
-	defer r.Close()
-
-	return newPackageDocument(r)
+	return getPackage(zr)
 }
 
 // GetPackageFromFile reads an epub's Open Package Document from an epub  file.
@@ -133,14 +128,29 @@ func GetPackageFromFile(path string) (*PackageDocument, error) {
 	return GetPackage(rzip)
 }
 
-func getContainer(rzip ReadAtSeeker) (*container, error) {
-	r, err := openFromZip(rzip, containerPath)
+func getContainer(zr *zip.Reader) (*container, error) {
+	r, err := zr.Open(containerPath)
 	if err != nil {
 		return nil, err
 	}
 	defer r.Close()
 
 	return newContainer(r)
+}
+
+func getPackage(zr *zip.Reader) (*PackageDocument, error) {
+	c, err := getContainer(zr)
+	if err != nil {
+		return nil, err
+	}
+
+	r, err := zr.Open(c.Rootfiles.FullPath)
+	if err != nil {
+		return nil, err
+	}
+	defer r.Close()
+
+	return newPackageDocument(r)
 }
 
 func getMeta(mdata *Metadata) *MetaInformation {