+++ ziputil.go
@@ -3,27 +3,12 @@ package epub
 import (
 	"archive/zip"
 	"io"
-	"os"
 )
 
 // ReadAtSeeker groups a io.ReaderAt and a io.Seeker.
 type ReadAtSeeker interface {
 	io.ReaderAt
 	io.Seeker
-}
-
-func openFromZip(r ReadAtSeeker, path string) (io.ReadCloser, error) {
-	zr, err := newZipReader(r)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, f := range zr.File {
-		if f.Name == path {
-			return f.Open()
-		}
-	}
-	return nil, os.ErrNotExist
 }
 
 func newZipReader(r ReadAtSeeker) (*zip.Reader, error) {